Remove multimixin API and make miin more flexible
Mixin can now take the
`{ mixin, initializer. aspectAdvice }` object structure,
or any object with a `factoryDescriptor` function. The
factoryDescriptor function is expected to return the same
object structure. By default now, `ComposeFactories` will
provide an implementaitonm of the `factoryDescriptor`
function. This implementation will return the factory as
the mixin, and will return _aspectAdvice or _initializer
properties as the aspectAdvice and initializer if they are
present on the factory prototype.