Merge arrays on prototypes
This commit allows for arrays to merged on prototypes when the target
and the source contain a property that is an array in both.  In addition
any arrays in the prototype during construction will be duplicated
so that the instance will not refer to the array in the prototype.

Fixes #46